layoutからheaderを分離
