+++ src/components/header.tsx
@@ -1,0 +1,33 @@
+import * as React from 'react';
+import Link from 'gatsby-link';
+
+const Header = () => (
+  <div
+    style={{
+      background: 'rebeccapurple',
+      marginBottom: '1.45rem',
+    }}
+  >
+    <div
+      style={{
+        margin: '0 auto',
+        maxWidth: 960,
+        padding: '1.45rem 1.0875rem',
+      }}
+    >
+      <h1 style={{ margin: 0 }}>
+        <Link
+          to="/"
+          style={{
+            color: 'white',
+            textDecoration: 'none',
+          }}
+        >
+          Gatsby
+        </Link>
+      </h1>
+    </div>
+  </div>
+);
+
+export default Header;
