+++ src/components/layout.tsx
@@ -1,38 +1,9 @@
-import * as React from 'react'
-import Link from 'gatsby-link'
-import Helmet from 'react-helmet'
-
-const Header = () => (
-  <div
-    style={{
-      background: 'rebeccapurple',
-      marginBottom: '1.45rem',
-    }}
-  >
-    <div
-      style={{
-        margin: '0 auto',
-        maxWidth: 960,
-        padding: '1.45rem 1.0875rem',
-      }}
-    >
-      <h1 style={{ margin: 0 }}>
-        <Link
-          to="/"
-          style={{
-            color: 'white',
-            textDecoration: 'none',
-          }}
-        >
-          Gatsby
-        </Link>
-      </h1>
-    </div>
-  </div>
-)
+import * as React from 'react';
+import Header from './header';
+import Helmet from 'react-helmet';
 
 interface DefaultLayoutProps extends React.HTMLProps<HTMLDivElement> {
-  children: any
+  children: any;
 }
 
 class DefaultLayout extends React.PureComponent<DefaultLayoutProps> {
@@ -58,8 +29,8 @@ class DefaultLayout extends React.PureComponent<DefaultLayoutProps> {
           {this.props.children}
         </div>
       </div>
-    )
+    );
   }
 }
 
-export default DefaultLayout
+export default DefaultLayout;
